upload so far
this has been a struggle so far, not sure if it's how the instructions are written but I seem to be making a lot of mistakes around element placement which have been fixed with trial/error.

it's given me some good practice about researching solutions on the help section, and also online so maybe it's by design?