project completed
this project was a struggle, i found a number of the steps unclear with what it was asking me to do, and there was a lot of trial and error because of this.

on the upside, it gave me plenty of opportunities to research solutions online and in the forum.