stories completed
this was a fairly underwhelming section, i met the story objectives but didn't feel it really added much to my overall understanding of CSS. I felt the steps in the previous module much more challenging and for now, I'm keen to learn the next segment, leaving this meeting it's objectives for now. I might come back later and add to it.