no fixed menu
after reading through stack overflow, it appears the fixed menu is using the position: fixed and then the left margin of the main would be pushed out by about 300px to compensate for the overflow.

while I -could- do this, it wasn't part of this project and I've still understood that fixed menus with CSS are possible, so long as there's margins there.

I might come back later and do it, however I want to continue with the course and not obsess over small features.

what's the saying? done is better than perfect?

changed some font colors too because black text isn't as readable.

experimented with the hover pseudo selector because I didn't like how the example version didn't make it noticeable it was clickable. i like the hover :)

this is  the final version before submitting the test :)