already in progress
started this on the train using an iPad so i didn't have access to github easily.

so far this has been interesting, i didn't realise variables could be used but I can see how this would be great for things like colour palettes or dark/light modes.

plus more practice with flexbox is a always welcome.