finished.
plenty of practice with variables and color changing. i didn't find this challenging, more tedious than anything. i think there's likely to be some good examples in here, but it was such a long segment, the fun wasn't there. it was a relief to finish it :)

the media query changing from night to day was a cool feature.