'technically' done
added section for the navbar which makes it fixed, following this: https://www.freecodecamp.org/news/how-to-keep-a-navbar-at-the-top-of-my-viewport/

what this doesn't mention is the rest of the page does need some padding/margin to begin otherwise it cuts off the top.

the current version of the page does fit the cert requirements, but it's ugly.

Before carrying on, I am going to spend some time making it look a bit nicer.